+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.26 (in progress)
 - Update GeoMx NGS directory schema
+- Update MERFISH directory schema
 
 ## v0.0.25
 - Update GeoMx NGS directory schema