+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Update GeoMx NGS directory schema
 - Update MERFISH directory schema
 - Update LC-MS directory schema
+- Updating tests
 
 ## v0.0.25
 - Update GeoMx NGS directory schema